Prevent footer from being too close to page header
This first attempt at Issue #4 follows the example given here:

    https://getbootstrap.com/examples/sticky-footer/

This isn't quite working right, but it's a start. The current problem is
that the footer's height ought to vary depending on the width of the
viewport; so the "magic number" of 180px in the CSS for the footer's
height sometimes looks too big and other times looks too small.