Update BibTeX for various papers
On TR-07-06, the name of the pdf was changed as well. I had changed the name of the pdf previously
because it was titled 07-09, even though it was for the 07-06 paper. This caused the index.md to
be incorrect for 07-06 because it was looking for the pdf titled 07-09.

For all other index.md files, the bibliography information was added.