Update the grants page
A few grants were added and a new feature of the grants page was added.
A grant can now be marked as primary or not for those grants where
multiple were awarded to various people under the same title. Professor
Rajan would like his to be marked primary and the others to be be not.
A nonprimary grant isn't displayed on the grants page.