Modify the papers sorting
The user can now select how the papers are sorted - either by project,
which has the type of the paper as a secondary sort, or by date, which
is the same sorting mechanism that was previously in place. The buttons
activate a javascript function that enables/ disables the visibility
of the different sorts.